+++ modules/app-android/src/androidMain/kotlin/org/archivekeep/app/android/AndroidEnvironment.kt
@@ -43,6 +43,6 @@ class AndroidEnvironment(
     override val storageDrivers =
         mapOf(
             "filesystem" to FileSystemStorageDriver(scope, fileStores),
-            "grpc" to GRPCStorageDriver(credentialsStore),
+            "grpc" to GRPCStorageDriver(scope, credentialsStore),
         )
 }