+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/desktop/ui/designsystem/sections/SectionTitle.kt
@@ -5,6 +5,8 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -18,6 +20,7 @@ import androidx.compose.ui.unit.sp
 @Composable
 fun SectionTitle(
     text: String,
+    isLoading: Boolean = false,
     buttons: @Composable RowScope.() -> Unit = {},
 ) {
     Row(
@@ -40,6 +43,9 @@ fun SectionTitle(
                     ),
                 modifier = Modifier.alignByBaseline(),
             )
+            if (isLoading) {
+                CircularProgressIndicator(Modifier.size(14.dp))
+            }
         }
 
         Row {