+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/desktop/ui/views/home/viewModel.kt
@@ -24,7 +24,6 @@ import org.archivekeep.app.desktop.domain.wiring.ArchiveOperationLaunchers
 import org.archivekeep.app.desktop.enableUnfinishedFeatures
 import org.archivekeep.app.desktop.utils.Action
 import org.archivekeep.utils.loading.Loadable
-import org.archivekeep.utils.loading.isLoading
 import org.archivekeep.utils.loading.mapIfLoadedOrDefault
 import org.archivekeep.utils.loading.mapLoadedData
 import org.archivekeep.utils.safeCombine